pyproject.toml
[project]
name = "ai-agent-resume-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-openai>=1.1.12",
    "pydantic-settings>=2.13.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "reportlab>=4.4.10",
    "uvicorn[standard]>=0.43.0",
]
